# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mocker admin API",
    "version": "1.0.0",
    "description": "The admin plane of mocker: everything the web UI talks to.\n\nThis document is the CONTRACT, not documentation-after-the-fact. web/ generates its\nentire API client from it (orval), and internal/admin/openapi_contract_test.go fails\nthe Go build's test bar if a single (method, path) here disagrees with the routes\ninternal/admin actually registers. Adding a handler without adding it here — or the\nreverse — is a red test, not a silent drift.\n\nEvery route below is served on the admin host (MOCKER_ADMIN_HOST) only. The mock\nplane (per-workspace hosts) is a different surface entirely and is NOT described here."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as the UI: the SPA is served by this very binary, so the client never needs an absolute base URL."
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mocker_session",
        "description": "Set by POST /api/auth/login. HttpOnly, SameSite=Lax, and Secure unless MOCKER_DEV=1."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Required on every POST/PUT/PATCH/DELETE except login. Value comes from the login/me response's csrfToken. Compared in constant time. Two further checks apply to the same requests and are NOT expressible here: Content-Type must parse to exactly application/json, and Origin (or Referer, when Origin is absent) must have hostname == MOCKER_ADMIN_HOST."
      }
    },
    "parameters": {
      "workspaceId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Workspace id.",
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request: unparseable body, an out-of-range field, an invalid path parameter.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "No session, or an expired one. The UI treats this on any non-login call as \"the session died\" and returns to the login screen.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The addressed resource does not exist.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Internal": {
        "description": "Server-side failure. The message is deliberately generic; the detail is in the server log.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ScenarioDetailView": {
        "type": "object",
        "description": "The full decoded snapshot — deliberately NOT the list shape (see ScenarioListView): this is what screen 5's active-scenario banner reads to learn which settings/overrides keys a running scenario currently masks. GET .../operations/{opKey} is never composed with an active scenario (A18), so this route is the only place that view is visible at all.",
        "required": [
          "id",
          "name",
          "createdAt",
          "isActive",
          "settings",
          "basePath",
          "spec",
          "overrides",
          "editVersion"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "description": "Unix seconds."
          },
          "isActive": {
            "type": "boolean"
          },
          "settings": {
            "$ref": "#/components/schemas/Settings"
          },
          "basePath": {
            "type": "string",
            "description": "Duplicates settings.basePath (the bundle format's own A14 rule). Never actually composed from a scenario snapshot when it is active — the WORKSPACE's own basePath is what a route table is always built from, so a scenario snapshotted at another prefix cannot silently relocate every route."
          },
          "spec": {
            "$ref": "#/components/schemas/ScenarioSpecRefView"
          },
          "overrides": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScenarioOverrideEntry"
            }
          },
          "editVersion": {
            "type": "integer",
            "format": "int64",
            "description": "A3's per-row compare-and-swap token for the scenarios row: the value to send back as PUT .../scenarios/{sid}'s editVersion. Both this read and that write's response share this schema, so a caller can write twice without re-reading between them."
          }
        }
      },
      "ScenarioListView": {
        "type": "object",
        "required": [
          "scenarios"
        ],
        "properties": {
          "scenarios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScenarioSummaryView"
            }
          }
        }
      },
...
      "CreateScenarioRequest": {
        "type": "object",
        "description": "Snapshots the workspace's CURRENT settings and op_overrides rows into a new named scenario (there is no scenario editor — a scenario's contents can only be produced by putting the workspace into that state and snapshotting it). Refused with 409 while another scenario is already active for this workspace and from is absent (deactivate it first, or pass from to clone instead).",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "from": {
            "type": "integer",
            "format": "int64",
            "description": "Id of a scenario already saved in this same workspace. When present, the new scenario's snapshot is a clone of THAT scenario's stored snapshot rather than a snapshot of the workspace's current layer, and its presence bypasses the 409 refusal that otherwise applies while a scenario is active for this workspace. The clone never reads the workspace's own layer, so the coherent-read retry (A11) does not apply to it either — a name collision with an existing scenario in this workspace still answers 409."
          }
        }
      },
...
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/workspaces/{id}/scenarios": {
      "parameters": [
        {
          "$ref": "#/components/parameters/workspaceId"
        }
      ],
      "get": {
        "tags": [
          "scenarios"
        ],
        "operationId": "listScenarios",
        "summary": "Scenarios saved for a workspace",
        "description": "Oldest first. Each entry is a name, a date and whether it is the active one — never the snapshot itself; fetch GET .../scenarios/{sid} for that.",
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioListView"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "post": {
        "tags": [
          "scenarios"
        ],
        "operationId": "createScenario",
        "summary": "Save the workspace's current state as a new scenario",
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScenarioRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDetailView"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "A scenario with this name already exists, a scenario is already active for this workspace and from was not given (deactivate it first, or pass from to clone instead — A10), or the workspace kept changing while it was being snapshotted and from was not given (retry — A11; a clone never reads the workspace's own layer, so this cause cannot fire when from is present).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
...
    "/api/workspaces/{id}/scenarios/{sid}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/workspaceId"
        },
        {
          "$ref": "#/components/parameters/scenarioId"
        }
      ],
      "get": {
        "tags": [
          "scenarios"
        ],
        "operationId": "getScenario",
        "summary": "One scenario's full snapshot",
        "description": "What screen 5's active-scenario banner reads to learn which settings/overrides keys a running scenario currently masks.",
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDetailView"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "No such scenario id in this workspace — including one that belongs to a DIFFERENT workspace (A8): the two answer identically, so this workspace never learns which is which.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "put": {
        "tags": [
          "scenarios"
        ],
        "operationId": "renameScenario",
        "summary": "Rename a scenario",
        "description": "Changes the scenario's name and nothing else — the stored snapshot, its created_at and whether it is the workspace's active scenario are untouched, and no revision is bumped (a rename does not change what any runtime serves). Renaming a scenario a running test suite names on the mock plane's own scenario-switch route is a breaking change for that suite.",
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameScenarioRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Renamed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDetailView"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "No such scenario id in this workspace (A8).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Two distinct meanings share this status on this route, told apart by error.code: conflict — a scenario with this name already exists in this workspace; or edit_conflict (A3) — editVersion no longer matches the row's stored edit_version, another write landed between this caller's read and this write, and details is either the scenario's current document (ScenarioConflictDetails) or, when the scenario was deleted meanwhile, the gone tombstone {\"gone\": true, \"editVersion\": null}. This response object is deliberately separate from the inline Error-typed 409 this route answered before A3.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioPutConflict"
                }
              }
            }
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "delete": {
        "tags": [
          "scenarios"
        ],
        "operationId": "deleteScenario",
        "summary": "Delete a scenario",
        "description": "Deleting the ACTIVE scenario deactivates it in the same write (ON DELETE SET NULL) and bumps the workspace's revision, so the next request stops being served from a runtime built over a scenario that no longer exists (A9).",
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "No such scenario id in this workspace (A8).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
...
  }
}